Update jbuschman_BBR2
Added the code for the sounds stuff. That code is currently not tested. Alot of the numbers in there are guesses or place holders for channels/pins. The dfplayer code I got from Terry Bauman and was modified to work for what I wanted. His code also had stuff for LEDs which I stripped out of this because I wont have any in my R2. At least not right now.